added queue name to the process
